# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "inventoryManagement",
    "contact": {}
  },
  "host": "54.176.169.179:3000",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/items/getItem": {
      "get": {
        "summary": "getItem",
        "tags": ["Products"],
        "operationId": "getItem",
        "deprecated": false,
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "authToken"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32",
            "description": "Optional",
            "default": 1
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32",
            "description": "Optional"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/items/updateItem": {
      "patch": {
        "summary": "updateItem",
        "tags": ["Products"],
        "operationId": "updateItem",
        "deprecated": false,
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "authToken"
          },
          {
            "name": "name",
            "in": "formData",
            "required": false,
            "type": "string",
            "description": "Optional"
          },
          {
            "name": "description",
            "in": "formData",
            "required": false,
            "type": "string",
            "description": "Optional"
          },
          {
            "name": "price",
            "in": "formData",
            "required": true,
            "type": "number",
            "format": "double",
            "description": "Optional"
          },
          {
            "name": "quantity",
            "in": "formData",
            "required": false,
            "type": "integer",
            "format": "int32",
            "description": "Optional"
          },
          {
            "name": "categoryId",
            "in": "formData",
            "required": false,
            "type": "string",
            "description": "Optional"
          },
          {
            "name": "itemId",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "Required"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/order/gettransactionHistory": {
      "get": {
        "summary": "get Transaction history",
        "tags": [
          "Order"
        ],
        "operationId": "getTransactionhistory",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/order/repeatOrder": {
      "post": {
        "summary": "repeat Order",
        "tags": [
          "Order"
        ],
        "operationId": "repeatOrder",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          },
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/repeatOrderRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/order/userOrder": {
      "get": {
        "summary": "userOrder",
        "tags": ["Order"],
        "operationId": "userOrder",
        "deprecated": false,
        "produces": ["application/json"],
        "consumes": ["application/x-www-form-urlencoded"],
        "parameters": [
        
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "authToken"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "repeatOrderRequest": {
      "title": "repeatOrderRequest",
      "example": {
        "orderId": "62df690b1766906a66007f98"
      },
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string"
        }
      },
      "required": [
        "orderId"
      ]
    },
...
  },
...
}